# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Price Encryption API",
    "description": "API to encrypt and decrypt prices.<h4>Additional Information</h4>",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Check server health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/encrypt": {
      "post": {
        "tags": ["Encryption"],
        "summary": "Encrypt a price",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "encryptionKey",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "integrityKey",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Encrypted price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncryptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Encryption failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/decrypt": {
      "post": {
        "tags": ["Decryption"],
        "summary": "Decrypt a price",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecryptRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "encryptionKey",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "integrityKey",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecryptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Decryption failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "OK" }
        }
      },
      "EncryptRequest": {
        "type": "object",
        "properties": {
          "price": { "type": "string", "example": "123.45" }
        },
        "required": ["price"]
      },
      "EncryptResponse": {
        "type": "object",
        "properties": {
          "encryptedPrice": { "type": "string", "example": "55mPuzTBIUqep2Kx_tebpjBGWi_5Y7jBFAOP" }
        }
      },
      "DecryptRequest": {
        "type": "object",
        "properties": {
          "encryptedPrice": { "type": "string", "example": "55mPuzTBIUqep2Kx_tebpjBGWi_5Y7jBFAOP" }
        },
        "required": ["encryptedPrice"]
      },
      "DecryptResponse": {
        "type": "object",
        "properties": {
          "decryptedPrice": { "type": "string", "example": "123.45" }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "nullable": true },
          "title": { "type": "string", "nullable": true },
          "status": { "type": "integer", "format": "int32", "nullable": true },
          "detail": { "type": "string", "nullable": true },
          "instance": { "type": "string", "nullable": true }
        },
        "additionalProperties": {}
      }
    }
  }
}